Map files and classes (#5)
* Testing commit with my branch

* Constructors and operators are written

Just finished the necessary 3 things we need for each class, implementation with source will follow

* More work on the constructors

Initialized and made some minor changes to the constructors

* More implementation to the source

More constructors should be working as intended and changes to the stream insertion operators.

* Major changes to the territories

Added a bool member so as to make a DFS traversal easier. I also added the DFS algorithm that traverses every neighbor and marks them as visited.

* Major rework to the map files

Had a lot of issues testing with pointers all over the place. I kept things consistent this time.

* Map files should be done

Minor adjustments to pointers and small tests with drivers

* Minor update to DFS with dangling pointers

* some updates

* Fixed a lot of the issues

Separated the classes back as before for convenience of the team and issue with inheritance. 
Improved DFS algo without pointers.
Renamed a lot of variables.
Map Driver is currently working fine with lots of debug statements.

* Fixes to edge connection

* Both drivers working now

* Added the debug variable

* Deleted lots of commented code and moved Driver